Enum ckb_verification::TransactionError

pub enum TransactionError {
    InsufficientCellCapacity,
    OutputsSumOverflow,
    Empty,
    DuplicateDeps,
    OutputsDataLengthMismatch,
    OutputDataHashMismatch,
    InvalidSince,
    Immature,
    CellbaseImmaturity,
    MismatchedVersion,
    ExceededMaximumBlockBytes,
}

Variants

InsufficientCellCapacity

output.occupied_capacity() > output.capacity()

OutputsSumOverflow

SUM([o.capacity for o in outputs]) > SUM([i.capacity for i in inputs])

Empty

inputs.is_empty() || outputs.is_empty()

DuplicateDeps

Duplicated dep-out-points within the same one transaction

OutputsDataLengthMismatch

outputs.len() != outputs_data.len()

OutputDataHashMismatch

ANY([o.data_hash != d.data_hash() for (o, d) in ZIP(outputs, outputs_data)])

InvalidSince

The format of transaction.since is invalid

Immature

The transaction is not mature which is required by transaction.since

CellbaseImmaturity

The transaction is not mature which is required by cellbase maturity rule

MismatchedVersion

The transaction version is mismatched with the system can hold

ExceededMaximumBlockBytes

The transaction size is too large

Methods

impl TransactionError

pub fn is_malformed_tx(&self) -> bool